% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line_RangePlot.R
\name{line_range_plot}
\alias{line_range_plot}
\title{Line range plot function with 'ggplot2'}
\usage{
line_range_plot(data, x_axis, y-min, y_max)
}
\arguments{
\item{data}{A data frame.}

\item{x_axis}{The column in the data frame to be plotted on the x-axis
of the graph.}

\item{y-min}{The column in the data frame to be plotted with the y-max data.}

\item{y-max}{The column in the data frame to be plotted with the y-min data.}
}
\value{
The line range plot of the selected data.
}
\description{
Function for plotting a line range function using the 'ggplot2' package.
}
\details{
This function includes an x-axis label rotation at -45 degrees for easy axis
reading.
}
\examples{
Line range plot using the 'change_na_rm' function to remove NA values with 
additional y-axis title labeling:

ants \%>\% 
 change_na_rm("None") \%>\% 
 line_range_plot(Tribe, min_ma, max_ma) +
 labs(y = "Age (in Ma)")

 
Line range plot using additional axis title labeling and color distinction:

butterfly \%>\% 
line_range_plot(ButterflySpecies, SpringTemp, SummerTemp) + 
aes(color = Day) +
labs(x = "Butterfly Species", 
     y = "Spring and Summer Temperature Range") 
}
