% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line_RangePlot.R
\name{line_range_plot}
\alias{line_range_plot}
\title{Function for plotting a line range function using the 'ggplot2' package.}
\usage{
Line range plot using the 'change_na_rm' function to remove NA values with additional y-axis labeling 
ants %>% 
 change_na_rm("None") %>% 
 line_range_plot(Tribe, min_ma, max_ma) +
 labs(y = "Age (in Ma)")
}
\arguments{
\item{data}{a data frame}

\item{x_axis}{the column in the data frame to be plotted on the x-axis of the graph}

\item{y-min}{the column in the data frame to be plotted with the y-max data}

\item{y-max}{the column in the data frame to be plotted with the y-min data}
}
\value{
plot The line range plot of the selected data
}
\description{
This function includes an x-axis label rotation at -45 degrees for easy axis
reading.
}
