% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumCatPredictors_lm.R
\name{lm_num_cat}
\alias{lm_num_cat}
\title{Linear model plot function with 'ggplot2'
Function for plotting a linear regression model using the 'ggplot2' package.
This function includes a colored trend line.}
\usage{
lm_num_cat(ants, min_ma, max_ma, subfamily)
}
\arguments{
\item{data}{A data frame.}

\item{response}{A column in the data frame to be used as a response variable.}

\item{num_predictor}{A column in the data frame containing numerical values
to be used as the predictor variable.}

\item{cat_predictor}{A column in the data frame containing categorical values
to be used as an additional predictor variable distinguished by color;
If left blank, the linear regression plot will be returned with only one
predictor value.}
}
\value{
A linear regression plot with a colored trend line.
}
\description{
Linear model plot function with 'ggplot2'
Function for plotting a linear regression model using the 'ggplot2' package.
This function includes a colored trend line.
}
\examples{
  ants \%>\% 
   change_na_rm("None") \%>\% 
   lm_num_cat(min_ma, max_ma)
  ants \%>\%
   change_na_rm("None") \%>\%
   lm_num_cat(min_ma, max_ma, Genus) +
   annotate("text", x = 42, y = 27, 
   label = "R^2 == 0.987", parse = TRUE, size = 4)
}
