% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxMa_Genus_lm.R
\name{lm_num_cat}
\alias{lm_num_cat}
\title{Function for plotting a linear regression model using the 'ggplot2' package.}
\usage{
no color/categorical predictor argument using the 'change_na_rm' function to remove NA values before plotting the linear regression model
ants %>% 
 change_na_rm("None") %>% 
 lm_num_cat(min_ma, max_ma)

color/categorical predictor argument using the 'change_na_rm' function to remove NA values with additional annotation for plot for the linear regression model
ants %>%
 change_na_rm("None") %>%
 lm_num_cat(min_ma, max_ma, Genus) +
 annotate("text", x = 42, y = 27, label = "R^2 == 0.987", parse = TRUE, size = 4)
}
\arguments{
\item{data}{a data frame}

\item{response}{a column in the data frame to be used as a response variable}

\item{num_predictor}{a column in the data frame containing numerical values
to be used as the predictor variable}

\item{cat_predictor}{a column in the data frame containing categorical values
to be used as an additional predictor variable distinguished by color
if left blank, the linear regression plot will be returned with only one
predictor value}
}
\value{
plot A linear regression plot with a colored trend line
}
\description{
This function includes a colored trend line.
}
