% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dropNA.R
\name{change_na_rm}
\alias{change_na_rm}
\title{Function for changing blank or non-NA values in a given data frame into NA
values if not yet changed to NA, then removing the NA values from the data
frame in one function.}
\usage{
change_na_rm(ants, "None")
change_na_rm(combined)
for removing NA values only
}
\arguments{
\item{data}{a data frame containing missing, non-NA, or NA values}

\item{characterName}{the non-NA or missing value to be replaced by NA
if this argument is left blank, the function will only
remove the NA values that are already present in the data frame}
}
\value{
data frame The new data frame with missing, non-NA, or NA values replaced (if applicable) and removed.
}
\description{
This function is beneficial for data frames being used with the 'tidyverse'
package, but can also be used in base R.
}
